rtw89: Make some changes in signing
The Makefile has been modified to incorporate the changes of user
hatrot3000 and some of mine. I build an efi system to check it out. At
least on openSUSE Tumbleweed, the modules build correctly.

The process needs a little more tweaking as there is no need to create
the sign info once it is created.

Signed-off-byL Larry Finger@lwfinger.net